Fix spacing, add SamplerKey.cs, remove Fighter jet.glb
