--- Fighter jet.glb
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4fc9d5574ec12fab57fd19d8c27fcbe60ff8ab40d2c32a0eb38f51973092e8a1
-size 42185260
