+++ Runtime/Scripts/SamplerKey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81a4096ec93ed3346b0ab71bc8359274
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
